requirements.txt
flask==2.2.3
pandas==1.5.3
numpy==1.24.2
openai==0.27.0
scikit-learn==1.2.2
hdbscan==0.8.29
umap-learn==0.5.3
matplotlib==3.5.1
youtube-comment-downloader==0.1.1